+++ CHANGELOG.rst
@@ -2,8 +2,8 @@
 CHANGELOG
 =========
 
-1.9.1dev
-========
+1.9.1
+=====
 
 * bug-fix: Estimators: Fix serialization of single records
 * bug-fix: deprecate enable_cloudwatch_metrics from Framework Estimators.